requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
pyserial==3.5
Pillow==10.0.0
